Move ORM/CSV code into `orm_factory`
We want to be able to use this outside of the CSVQueryEngine.

We also remove the undocumented `create_if_missing` option because we're
going to take a different approach here.
